back/requirements.txt
flask==3.1.*
gunicorn==21.*
twilio==9.*
pydub==0.25.*
google-genai>=1.0,<2.0
langdetect==1.0.*
requests==2.32.*
python-dotenv==1.0.*
gTTS==2.5.*
elevenlabs>=2.0,<3.0
pydantic>=2.0,<3.0
redis>=5.0,<6.0
sqlalchemy>=2.0,<3.0
psycopg2-binary>=2.9,<3.0
pgvector>=0.3,<1.0
flask-cors>=5.0,<7.0
